+++ android/MobileSdk/src/main/java/com/spruceid/mobile/sdk/ble/BleSecurityUtils.kt
@@ -156,147 +156,4 @@ object BleSecurityUtils {
             logger.d("Random delay interrupted")
         }
     }
-
-    /**
-     * Generate cryptographically secure random bytes
-     *
-     * @param size Number of bytes to generate
-     * @return Secure random byte array
-     */
-    fun generateSecureRandom(size: Int): ByteArray {
-        if (size <= 0 || size > 1024) {
-            throw IllegalArgumentException("Invalid random size: $size (must be 1-1024)")
-        }
-
-        return ByteArray(size).apply {
-            secureRandom.nextBytes(this)
-        }
-    }
-
-    /**
-     * Securely clear sensitive data from memory
-     *
-     * Overwrites byte arrays containing sensitive data (keys, identifiers)
-     * to prevent memory-based attacks or accidental disclosure.
-     *
-     * @param sensitiveData Array to clear
-     */
-    fun secureErase(sensitiveData: ByteArray?) {
-        sensitiveData?.let { data ->
-            // Overwrite with random data first
-            secureRandom.nextBytes(data)
-            // Then overwrite with zeros
-            data.fill(0)
-
-            logger.d("Securely erased ${data.size} bytes")
-        }
-    }
-
-    /**
-     * Validate message integrity using simple checksum
-     *
-     * Provides basic integrity checking for BLE messages to detect
-     * corruption during transmission. Not cryptographically secure,
-     * but sufficient for detecting accidental corruption.
-     *
-     * @param data Message data
-     * @return Simple checksum value
-     */
-    fun calculateSimpleChecksum(data: ByteArray): Int {
-        var checksum = 0
-        for (byte in data) {
-            checksum = (checksum + (byte.toInt() and 0xFF)) and 0xFFFF
-        }
-        return checksum
-    }
-
-    /**
-     * Validate BLE security requirements
-     *
-     * Checks if the current BLE connection meets minimum security
-     * requirements for mDL transactions per ISO 18013-5.
-     *
-     * @param securityLevel Current BLE security level
-     * @param requireEncryption Whether encryption is mandatory
-     * @param requireAuthentication Whether authentication is mandatory
-     * @return true if security requirements are met
-     */
-    fun validateBleSecurityLevel(
-        securityLevel: Int,
-        requireEncryption: Boolean = true,
-        requireAuthentication: Boolean = true
-    ): Boolean {
-        // These constants would typically come from BluetoothDevice or BluetoothGatt
-        val hasEncryption = securityLevel >= 2  // Placeholder - use actual constants
-        val hasAuthentication = securityLevel >= 3  // Placeholder - use actual constants
-
-        val securityOk = (!requireEncryption || hasEncryption) &&
-                        (!requireAuthentication || hasAuthentication)
-
-        if (!securityOk) {
-            logger.w("BLE security requirements not met: level=$securityLevel, " +
-                    "encryption=$hasEncryption, authentication=$hasAuthentication")
-        }
-
-        return securityOk
-    }
-
-    /**
-     * Sanitize log data to prevent sensitive information leakage
-     *
-     * Removes or masks sensitive information from data before logging.
-     * Used to prevent accidental disclosure of mDL data, keys, or
-     * personal information in log files.
-     *
-     * @param data Raw data to sanitize
-     * @param maskLength Number of characters to show before masking
-     * @return Sanitized string safe for logging
-     */
-    fun sanitizeForLogging(data: ByteArray?, maskLength: Int = 4): String {
-        return when {
-            data == null -> "null"
-            data.isEmpty() -> "empty"
-            data.size <= maskLength -> "[MASKED:${data.size}]"
-            else -> {
-                val prefix = data.take(maskLength)
-                    .joinToString("") { "%02x".format(it) }
-                "$prefix...[MASKED:${data.size-maskLength}]"
-            }
-        }
-    }
-
-    /**
-     * Security configuration validation
-     *
-     * Validates that BLE configuration meets security requirements
-     * for production mDL deployments.
-     *
-     * @param config BLE configuration to validate
-     * @return List of security warnings/errors
-     */
-    fun validateSecurityConfiguration(config: BleConfiguration): List<String> {
-        val issues = mutableListOf<String>()
-
-        if (!config.randomizeResponseTiming) {
-            issues.add("Timing attack protection disabled")
-        }
-
-        if (config.logSensitiveData) {
-            issues.add("Sensitive data logging enabled (not recommended for production)")
-        }
-
-        if (config.connectionTimeoutMs > 30000) {
-            issues.add("Connection timeout too long (DoS risk)")
-        }
-
-        if (config.maxConnectionRetries > 10) {
-            issues.add("Too many retry attempts configured (DoS risk)")
-        }
-
-        if (config.messageTimeoutMs > 60000) {
-            issues.add("Message timeout too long (resource exhaustion risk)")
-        }
-
-        return issues
-    }
 }
